use java.time for the brute force test of the library as java.calendar itself is buggy. complete explanation below:
We were using java.calendar to test the library, but the library ended up testing java.calendar!
java.calendar had 2 severe bugs:
GregorianCalendar would consider years like 700 or 500 a leap year. which are not leap years in gregorian calendar but are in julian calendar.
GregorianCalendar would move from 1582/10/04 to 1582/10/15. 4th october was in julian calendar, and then they switched to GregorianCalendar. so gregorian calendar does have 1582/10/05 till 10/15.

also the conversion of milliseconds since epoch by this library was wrong due to the bugs above.